finishing delete comment use case + adding a test in comment repo

+++ src/Applications/use_case/_test/DeleteCommentUseCase.test.js
@@ -4,14 +4,17 @@ const DeleteCommentUseCase = require('../DeleteCommentUseCase');
 
 describe('DeleteCommentUseCase', () => {
   it('should orchestrating the delete comment action correctly', async () => {
-    const mockCommentRepository = new CommentRepository();
-    const mockThreadRepository = new ThreadRepository();
     const useCasePayload = {
       threadId: 'thread-321',
       commentId: 'comment-123',
       owner: 'user-123',
     };
 
+    const mockThreadRepository = new ThreadRepository();
+    const mockCommentRepository = new CommentRepository();
+
+    mockThreadRepository.findThreadById = jest.fn().mockImplementation(() => Promise.resolve());
+    mockCommentRepository.verifyCommentAccess = jest.fn().mockImplementation(() => Promise.resolve());
     mockCommentRepository.deleteComment = jest.fn().mockImplementation(() => Promise.resolve());
 
     const deleteCommentUseCase = new DeleteCommentUseCase({
@@ -23,6 +26,6 @@ describe('DeleteCommentUseCase', () => {
 
     expect(mockThreadRepository.findThreadById).toBeCalledWith(useCasePayload.threadId);
     expect(mockCommentRepository.verifyCommentAccess).toBeCalledWith(useCasePayload.commentId, useCasePayload.owner);
-    expect(mockCommentRepository.deleteComment).toBeCalledWith(useCasePayload.owner);
+    expect(mockCommentRepository.deleteComment).toBeCalledWith(useCasePayload.commentId);
   });
 });
